Refactor recipe duration formatter
