+++ common/package.json
@@ -1,8 +1,6 @@
 {
   "name": "common",
   "version": "1.0.0",
-  "dependencies": {
-  },
   "devDependencies": {
     "@tiptap/core": "^2.2.4"
   }
